Fix: more accurate tunnel duration calculation * Don't use current time as tunnel end time, since device resume   after sleep would include sleep time in duration * Instead, use last data received time as estimated end time
